% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular.R
\name{as.data.frame.supreme}
\alias{as.data.frame.supreme}
\title{Turn supreme data into a data frame}
\usage{
\method{as.data.frame}{supreme}(x, ...)
}
\arguments{
\item{x}{a \code{supreme} object.}

\item{...}{additional arguments to be passed to or from methods.}
}
\description{
Turn supreme data into a data frame
}
\examples{
\dontrun{
paths <- supreme_example()
Object <- supreme(src_file(paths))
as.data.frame(Object)
}
}
