% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classes.R
\name{classes}
\alias{classes}
\title{Class labels}
\description{
Labels that are assigned to cases when they come to the Supreme Court from lower courts.
We consider only the first 10 larger classes.
}
\details{
\describe{
  \item{class 71}{Tributi (taxes)}
  \item{class 42}{Lavoro (employment)}
  \item{class 30}{Equa riparazione (equitable relief)}
  \item{class 16}{Contratti: tutti gli altri tipi (contracts: other types)}
  \item{class 15}{Contratti e obbligazioni in genere (contracts and obligations in general)}
  \item{class 57}{Responsabilita' civile (civil liability)}
  \item{class 49 }{Previdenza (social security)}
  \item{class 19}{Diritti reali (property rights)}
  \item{class 34}{Fallimento e istituti affini (bankruptcy)}
  \item{class 74}{Vendita, Permuta, Riporto (sale, exchange, carry-over)}
}
}

