% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/corpus.R
\name{corpus}
\alias{corpus}
\title{A corpus of Italian Supreme Court decisions}
\description{
\code{corpus} is an object of class \code{VCorpus} from package \pkg{tm} and contains 15,485 documents
relating to final decisions (sentences) in civil matters delivered by the Italian Supreme Court during
the year 2013.
\code{\link{classes}} provide subject labels for each document.
}
\note{
\code{corpus} inherits user defined meta attibutes from \code{\link{dfciv}} columns.
In particular, each document is labeled by the \code{Idmateria} attribute.
}

