% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/corpus2dtm.R
\name{corpus2dtm}
\alias{corpus2dtm}
\title{From ISC corpus to a Document-Term Matrix}
\usage{
corpus2dtm(corpus, stopwords)
}
\arguments{
\item{corpus}{a corpus of decisions from Italian Supreme Court.}

\item{stopwords}{a character vector of stopwords.}
}
\value{
\code{dtm} a \emph{raw} document-term matrix with minimum term length 3 and terms appearing at least in 2 documents.
}
\description{
\code{corpus2dtm} transforms a corpus of decisions from Italian Supreme Court to a document term matrix.
}
\note{
Basic text cleaning steps build a \emph{raw} \code{dtm} by selecting only terms (columns)
corresponding to a suitable vocabulary. Typically, this involves converting tokens to lower-case,
removing punctuation characters, removing numbers, stemming, removing stop-words and selecting terms
with a length above a certain minimum and occurring at least in a minimum number of documents.
Package \pkg{tm version >= 0.6} required.
}
\examples{
\dontrun{
library(Supreme)
data("corpus")
data("italianStopWords")  # for removing italian stop words
dtm <- corpus2dtm(corpus, italianStopWords)
}
}

