% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dfciv2corpus.R
\name{dfciv2corpus}
\alias{dfciv2corpus}
\title{Creating an ISC corpus from a data frame of class dfciv}
\usage{
dfciv2corpus(dfciv, export = FALSE)
}
\arguments{
\item{dfciv}{a data frame of class \code{dfciv}.}

\item{export}{logical, if \code{TRUE} exports the corpus to txt files. Default is \code{FALSE}.}
}
\value{
a ISC \code{corpus} built from a \code{dfciv} data frame with user defined meta attributes (see \code{\link{xml2dfciv}} function).
}
\description{
\code{dfciv2corpus} transforms a data frame of class \code{dfciv} to a corpus of Italian Supreme Court
decisions and exports the result to the built-in directory \code{data/corpus}.
}
\examples{
\dontrun{
library(Supreme)
data("dfciv")
corpus <- dfciv2corpus(dfciv, TRUE)
}
}

