% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_environment}
\alias{example_environment}
\title{Get environment to \code{supreme} example}
\usage{
example_environment()
}
\description{
Get environment to \code{supreme} example
}
\examples{
env <- example_environment()
supreme(src_env(env))
}
\seealso{
Other source examples: \code{\link{example_app_path}},
  \code{\link{example_expression}},
  \code{\link{example_package}}
}
\concept{source examples}
