% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_expression}
\alias{example_expression}
\title{Get expression to \code{supreme} example}
\usage{
example_expression()
}
\description{
Get expression to \code{supreme} example
}
\examples{
expr <- example_expression()
supreme(src_expr(expr))
}
\seealso{
Other source examples: \code{\link{example_app_path}},
  \code{\link{example_environment}}
}
\concept{source examples}
