% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_package}
\alias{example_package}
\title{Get package to \code{supreme} example}
\usage{
example_package(pkg = list(name = "supreme.pkg.test", path =
  "strboul/supreme.pkg.test"))
}
\arguments{
\item{pkg}{a list contains name (\emph{package name}) and path (\emph{GitHub path}) fields.
For the example purposes, the default value is \code{supreme.pkg.test}.}
}
\description{
This function asks to build packages in the users' system if they are not found.
}
