% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_yaml}
\alias{example_yaml}
\title{Get YAML to \code{supreme} example}
\usage{
example_yaml()
}
\description{
Get YAML to \code{supreme} example
}
\details{
Please note that all examples give the same output.
}
\examples{
yaml <- example_yaml()
supreme(src_yaml(env))
}
\seealso{
Other source examples: 
\code{\link{example_app_path}()}
}
\concept{source examples}
