% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph}
\alias{graph}
\title{Make a graph of modules}
\usage{
graph(x)
}
\arguments{
\item{x}{a \code{supreme} object.}
}
\description{
The call graph creates a UML like graph of your Shiny application.
}
\examples{
\dontrun{
path <- example_app_path()
supr <- supreme(src_file(path))
graph(supr)
}
}
