% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mcLDA.R
\name{mcLDA}
\alias{mcLDA}
\title{Multicore parallel runs of LDA models}
\usage{
mcLDA(dtm, lda.method = c("VEM", "VEM_fixed", "Gibbs"), k.runs = list(from =
  2, to = 2, steps = 1), target)
}
\arguments{
\item{dtm}{a document-term matrix.}

\item{lda.method}{character. Approximate posterior inference method.}

\item{k.runs}{the grid of \code{k} values.}

\item{target}{factor. The labeling variable for logistic classification.}
}
\value{
a list containing the fitted LDA models and the misclassification errors.
Model with minimum misclassification error, i.e. the \emph{best} model, is also returned.
}
\description{
\code{mcLDA} runs multiple parallel LDA models by varying the number of topics \code{k} over a predefined grid of values
and performs model selection based on logistic classification by calling \code{\link{logClass}} function.
}
\details{
This function runs multiple LDA models and applies internal \code{\link{logClass}} to each model.
A vector of misclassification error on the test set (\code{e1.test}) is returned and the \emph{best} model
is selected with minimum misclassification error.
}
\note{
By default the \pkg{doParallel} package uses snow-like functionality. The snow-like functionality should work fine on Unix-like systems.
Output is automatically saved in directory \code{data/ws/output} and a log file is provided in directory \code{log}.
}
\examples{
\dontrun{
library(Supreme)
data("dtm")
data("classes")
dtm.lognet <- reduce_dtm(dtm, method = "lognet", target = classes)

# 4 cores: one model for each core.
mc.lda.models <- mcLDA(dtm.lognet$reduced, lda.method = "VEM", k.runs = list(from = 10, to = 25, steps = 5), target = classes)
}
}

