% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supreme.R
\name{is_list}
\alias{is_list}
\alias{is_expression}
\title{Checks if an object is a list (but not a data.frame)}
\usage{
is_list(x)

is_expression(x)
}
\description{
Checks if an object is a list (but not a data.frame)
}
