% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reduce_dtm_lognet_cv.R
\name{reduce_dtm_lognet_cv}
\alias{reduce_dtm_lognet_cv}
\title{Internal Supreme function}
\usage{
reduce_dtm_lognet_cv(dtm, classes, lambda = c("lambda.min", "lambda.1se"),
  SEED, c_normalize = TRUE, parallel = TRUE, export = FALSE)
}
\arguments{
\item{dtm}{a document-term matrix in term frequency format.}

\item{classes}{factor, the labeling variable.}

\item{lambda}{a string with the selection rule of the optimal fit.}

\item{SEED}{integer, the random seed for selecting train and test sets.}

\item{c_normalize}{logical. If \code{TRUE} \code{dtm} entries are (cosine) normalized. Default is \code{TRUE}.}

\item{parallel}{logical. If \code{TRUE} parallel cross-validation is performed. Default is \code{TRUE}.}

\item{export}{logical. If \code{TRUE} export the discarded terms, the vocabulary and the returned object
to the built-in directory \code{data/dtm}. Default is \code{FALSE}.}
}
\value{
a list with the \emph{reduced} \code{dtm} (in term frequency format),
the IDs of documents belonging to the training set, the \code{glmnet} fit object,
the position of the best lambda, the selected terms by class, and the train and test
misclassification errors \code{err1.train} and \code{err1.test}.
Confusion matrix is also returned.
}
\description{
\code{reduce_dtm_lognet_cv} reduces the number of terms (columns) of a \emph{labeled} document-term matrix.
\code{reduce_dtm_lognet_cv} is called by the \code{\link{reduce_dtm}} function.
}
\details{
This function fits a logistic classification model via penalized maximum likelihood
by calling the \code{lognet} function from package \pkg{glmnet}.
The regularization path is only computed for the \strong{lasso} penalty at a grid of values
for the regularization parameter \code{lambda}.
If \code{c_normalize = TRUE} (default) the \code{dtm} is passed
for cosine normalization to the \code{\link{wTfIdf}} function.
Reduction of number of terms is performed by selecting only columns corresponding
to the \strong{non zero} \emph{beta} coefficients in the optimal fit.
}
\note{
Tuning parameters \code{alpha} and \code{lambda} are respectively set in the optimal fit
to \code{1} (default) and one out of \code{lambda.min} or \code{lambda.1se}.
The latter follows from the "minimum training error rule" and the former
from the more conservative approach of the "one standard error rule".
Full details are given in \emph{"The Elements of Statistical Learnings"}
(T. Hastie, R. Tibshirani, J. Friedman) 2nd edition p. 61.
}
\examples{
\dontrun{
library(Supreme)
data("dtm")
data("classes")
dtm.lognet.cv <- reduce_dtm_lognet_cv(dtm, classes, lambda = "lambda.1se", SEED = 123)
}
}

