% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reduce_dtm_tfidf.R
\name{reduce_dtm_tfidf}
\alias{reduce_dtm_tfidf}
\title{Internal Supreme function}
\usage{
reduce_dtm_tfidf(dtm, q = list(inf = 0.25, sup = 0.75), export = FALSE)
}
\arguments{
\item{dtm}{a document-term matrix.}

\item{q}{a list with \code{inf} and \code{sup} quantiles of tf-idf scores distribution. Default are the first and third quartiles.}

\item{export}{logical. If \code{TRUE} exports the discarded terms, the vocabulary and the returned object to the built-in directory \code{data/dtm}. Default is \code{FALSE}.}
}
\value{
a list with the \emph{reduced} \code{dtm} and associated \code{term_tfidf}.
Values of quantile thresholds are also returned.
}
\description{
\code{reduce_dtm_tfidf} selects suitable columns of an \emph{unlabeled} document-term matrix
by deleting terms with tf-idf score out of a range defined by \code{inf} and \code{sup}
quantiles of tf-idf scores distribution.
\code{reduce_dtm_tfidf} is called by \code{\link{reduce_dtm}} function.
}
\examples{
\dontrun{
library(Supreme)
data("dtm")
dtm.tfidf <- reduce_dtm_tfidf(dtm, export = TRUE)
}
}

