% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src-methods.R
\name{src_env}
\alias{src_env}
\title{Read \R environment}
\usage{
src_env(x)
}
\arguments{
\item{x}{an \R environment.}
}
\description{
Read \R environment
}
\seealso{
Other source functions: \code{\link{src_expr}},
  \code{\link{src_file}}, \code{\link{src_pkg}},
  \code{\link{src_yaml}}
}
\concept{source functions}
