% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src-methods.R
\name{src_expr}
\alias{src_expr}
\title{Read \R expressions}
\usage{
src_expr(x)
}
\arguments{
\item{x}{an \R expression.}
}
\description{
Read \R expressions
}
\seealso{
Other source functions: 
\code{\link{src_env}()},
\code{\link{src_file}()},
\code{\link{src_pkg}()},
\code{\link{src_yaml}()}
}
\concept{source functions}
