% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src-methods.R
\name{src_file}
\alias{src_file}
\title{Read \R files}
\usage{
src_file(x)
}
\arguments{
\item{x}{a file path.}
}
\value{
A \code{src_file} object.
}
\description{
Read files contain at least one Shiny application.
}
\examples{
paths <- example_app_path()
s <- supreme(src_file(paths))
}
\seealso{
Other source functions: 
\code{\link{src_yaml}()}
}
\concept{source functions}
