% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src-methods.R
\name{src_pkg}
\alias{src_pkg}
\title{Read an \R package}
\usage{
src_pkg(x)
}
\arguments{
\item{x}{a package name as character (must be installed in the system).}
}
\description{
The package should contain some Shiny application.
}
\details{
The function coerces input to be a character.
}
\seealso{
Other source functions: 
\code{\link{src_env}()},
\code{\link{src_expr}()},
\code{\link{src_file}()},
\code{\link{src_yaml}()}
}
\concept{source functions}
