% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src-yaml.R
\name{src_yaml}
\alias{src_yaml}
\title{Read yaml model}
\usage{
src_yaml(file = NULL, text = NULL)
}
\arguments{
\item{file}{file path to a yaml file.}

\item{text}{a YAML formatted character string.}
}
\description{
Read yaml model
}
\examples{
\dontrun{
file <- file.path("yaml-test", "example-model-1.yaml")
src_yaml(file)

model <- "
- type: module
  name: childModuleA
  input: [input.data, reactive]
  calling_modules: grandChildModule1

- type: module
  name: childModuleB
  input: selected.model
"
src_yaml(text = model)
}
}
