% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supreme.R
\name{supreme}
\alias{supreme}
\title{Create a supreme object}
\usage{
supreme(x, expr = NULL)
}
\arguments{
\item{x}{file name(s) containing valid Shiny application(s).}

\item{expr}{an \R expression.}
}
\description{
Create a supreme object
}
