% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{supreme_example}
\alias{supreme_example}
\title{Get paths to \code{supreme} example}
\usage{
supreme_example(file = NULL)
}
\arguments{
\item{file}{file names. If no file names are put (which \code{path} is \code{NULL}), then
all the example file paths will be listed.}
}
\description{
Contains the paths of the Shiny application that is a fully-fledged example
endeavors to demonstrate all the features \code{supreme} has.
}
\examples{
supreme_example()
supreme_example(c("app.R", "module-customers.R"))
}
