% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supreme.R
\name{tree_app}
\alias{tree_app}
\title{Create a module tree}
\usage{
tree_app(x)
}
\arguments{
\item{x}{an \R expression or a file name that contain a (valid) Shiny application.}
}
\description{
Create a module tree
}
