% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wTfIdf.R
\name{wTfIdf}
\alias{wTfIdf}
\title{Internal Supreme function}
\usage{
wTfIdf(dtm, normalize = TRUE)
}
\arguments{
\item{dtm}{a document-term matrix in term frequency format.}

\item{normalize}{a Boolean value indicating whether the tf-idf scores should be normalized. Default is \code{TRUE}.}
}
\value{
The weighted document-term matrix.
}
\description{
Weight a document-term matrix by term frequency - inverse document frequency.
}
\details{
In order for the weights to fall in the \code{[0, 1]} interval and
for the documents to be represented by vectors of equal length,
the weights are normalized by cosine normalization.
}
\note{
This function is slightly different from \link{weightTfIdf} in the package \pkg{tm}
because of the different normalization approach.
}
\examples{
\dontrun{
library(Supreme)
data("dtm")
dtm_normalized <- wTfIdf(dtm, normalize = TRUE)
}
}

