% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xmlciv2df.R
\name{xmlciv2df}
\alias{xmlciv2df}
\title{From xml docs to dfciv data frame}
\usage{
xmlciv2df(path)
}
\arguments{
\item{path}{character string containing the path of xml input documents.}
}
\value{
\code{dfciv} a data frame of class \code{dfciv} with added columns \emph{Id_doc} and \emph{Idmateria} (Id subject of appealed decision).
}
\description{
\code{xmlciv2df} transforms the original xml documents (decisions in civil matters from Italian Supreme Court) to a \code{dfciv} data frame.
}
\details{
Each original document has a \strong{xml} tree structure starting at \code{snintegrale} top element
and branching to \code{testo} and \code{dispositivo} leave elements, respectively containing
the content of the document body and the case decision.
Document attributes are:
\describe{
  \item{\code{tipoProv}}{type of decision (D, I, O, S) - S stands for *Sentenza* (final decision)}
  \item{\code{annoDec}}{year of decision}
  \item{\code{numDec}}{number of decision}
  \item{\code{numSez}}{number of ISC section}
  \item{\code{annoNrgSic}}{year of filing lawsuit}
  \item{\code{nrgSic}}{number of filing lawsuit}
  \item{\code{annoProvOrig}}{year of appealed decision}
  \item{\code{numProvOrig}}{number of appealed decision}
  \item{\code{autorita}}{authority of appealed decision}
  \item{\code{localita}}{location of appealed decision}
  \item{\code{materia}}{subject of appealed decision}
}
}
\note{
\code{xmlciv2df} transforms the xml tree to a \code{list} which in turn is transformed to a \code{data frame}.
You can obtain the \code{dfciv} data frame loading \href{https://www.dropbox.com/s/t0eiqagdyd67ei7/ISC_Civil_2013.zip?dl=0}{this}
compressed xml file and then running the code in the example.
Data frame \code{dfciv} contains only final decisions belonging to the 10 larger \emph{Idmateria} classes.
}
\examples{
\dontrun{
library(Supreme)

# Unzip and load in memory the xml input file.
xml_input <- unzip("../ISC_Civil_2013.zip")
dfciv_original <- xmlciv2df(xml_input)

# Select the subset of final decisions.
dfciv_final <- subset(dfciv_original, tipoProv == "S")

# Select only the 10 larger classes.
cl <- names(sort(table(dfciv_final$Idmateria), decreasing = TRUE))[1:10]
dfciv <- subset(dfciv_final, Idmateria \%in\% cl)

str(dfciv)

rm(dfciv_original, dfciv_final)
}
}

