% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular.R
\name{as.data.frame.supreme}
\alias{as.data.frame.supreme}
\title{Turn supreme data into a data frame}
\usage{
\method{as.data.frame}{supreme}(x)
}
\arguments{
\item{x}{a \code{supreme} object.}
}
\description{
Turn supreme data into a data frame
}
\examples{
paths <- example_app_path()
Object <- supreme(src_file(paths))
as.data.frame(Object)
}
