% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular.R
\name{as.data.frame.supreme}
\alias{as.data.frame.supreme}
\title{Turn supreme data into a \code{data.frame}}
\usage{
\method{as.data.frame}{supreme}(x, ...)
}
\arguments{
\item{x}{a \code{supreme} object.}

\item{...}{methods to be passed onto.}
}
\value{
a \code{data.frame}.
}
\description{
Turn supreme data into a \code{data.frame}
}
\examples{
paths <- example_app_path()
sp <- supreme(src_file(paths))
as.data.frame(sp)
}
