% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_app_path}
\alias{example_app_path}
\title{Get paths to \code{supreme} example}
\usage{
example_app_path(file = NULL)
}
\arguments{
\item{file}{file names. If no file names are put (which \code{path} is \code{NULL}), then
all the example file paths will be listed.}
}
\description{
Contains the paths of the Shiny application that is a fully-fledged example
endeavors to demonstrate all the features \code{supreme} has.
}
\examples{
example_app_path()
example_app_path(c("app.R", "module-customers.R"))
}
