% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_app_path}
\alias{example_app_path}
\title{Get paths to supreme example}
\usage{
example_app_path(file = NULL)
}
\arguments{
\item{file}{file names. If no file names are put (which \code{path} is \code{NULL}), then
all the example file paths will be listed.}
}
\value{
a character vector containing the R file path for the example.
}
\description{
The example Shiny application to demonstrate all the capabilities of what
\code{supreme} offers.
}
\examples{
files <- example_app_path(c("app", "module-customers"))
supreme(src_file(files))
}
\seealso{
Other source examples: 
\code{\link{example_yaml}()}
}
\concept{source examples}
