% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_package}
\alias{example_package}
\title{Get package to \code{supreme} example}
\usage{
example_package(pkg = list(name = "supreme.pkg.test", path =
  "strboul/supreme.pkg.test"))
}
\arguments{
\item{pkg}{a list contains name (\emph{package name}) and path (\emph{GitHub path}) fields.
For the example purposes, the default value is \code{supreme.pkg.test}.}
}
\description{
This function asks user to install the example \emph{supreme} package only if the
package is not found in any of the known directories.
}
\details{
This supreme example helper function \code{example_package()} contains an R package named
\code{supreme.pkg.test} placed in this link:
\url{https://github.com/strboul/supreme.pkg.test}

This function uses
\link[devtools:install_github]{devtools::install_github} call to clone, build and install the package.
}
\examples{
\dontrun{
pkg <- example_package()
supreme(src_pkg(pkg))
}
}
\seealso{
Other source examples: \code{\link{example_app_path}},
  \code{\link{example_environment}},
  \code{\link{example_expression}}
}
\concept{source examples}
