% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_yaml}
\alias{example_yaml}
\title{Get YAML to supreme example}
\usage{
example_yaml()
}
\value{
a character vector containing the YAML file path for the example.
}
\description{
Get YAML to supreme example
}
\examples{
yaml <- example_yaml()
supreme(src_yaml(yaml))
}
\seealso{
Other source examples: 
\code{\link{example_app_path}()}
}
\concept{source examples}
