% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supreme.R
\name{module_tree}
\alias{module_tree}
\title{Create a module tree}
\usage{
module_tree(x)
}
\arguments{
\item{x}{an \R expression or a file name that contain a (valid) Shiny application.}
}
\description{
Create a module tree
}
