% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src-methods.R
\name{src_yaml}
\alias{src_yaml}
\title{Read a YAML file containing model}
\usage{
src_yaml(file = NULL, text = NULL)
}
\arguments{
\item{file}{file path to a YAML file.}

\item{text}{a YAML formatted character string.}
}
\description{
Reads an object or a file in YAML format and returns a model YAML object.
}
\examples{
\dontrun{
## Read from file:
file <- file.path("yaml-test", "example-model.yaml")
src_yaml(file)

## Read from text object:
model <- "
- name: childModuleA
  input: [input.data, reactive]
  calling_modules: grandChildModule1

- name: childModuleB
  input: selected.model
"
src_yaml(text = model)
}
}
\seealso{
Other source functions: \code{\link{src_env}},
  \code{\link{src_expr}}, \code{\link{src_file}},
  \code{\link{src_pkg}}
}
\concept{source functions}
