% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supreme.R
\name{supreme}
\alias{supreme}
\title{Create a supreme object}
\usage{
supreme(x)
}
\arguments{
\item{x}{a valid source input.}
}
\value{
a \code{supreme} object.
}
\description{
Create a supreme object
}
\examples{
path <- example_app_path()
supreme(src_file(path))
}
